fix: combine share-type into events
