+++ packages/shared-type/src/index.ts
@@ -1,6 +1,5 @@
 export * from './libs/attendee';
 export * from './libs/event';
-export * from './libs/leaderboard';
 export * from './libs/organizer';
 export * from './libs/post';
 export * from './libs/user';
