+++ packages/shared-type/src/libs/event.ts
@@ -39,6 +39,26 @@ export const eventInvResponse = z.object({
   accepted: z.boolean(),
 });
 
+export const leaderboardEntry = z.object({
+  name: z.string(),
+  uid: z.string(),
+  eid: z.string(),
+  score: z.number().nullable(),
+});
+
+export const leaderboardData = z.array(leaderboardEntry);
+
+export const createLeaderboardEntry = z.object({
+  name: z.string(),
+  uid: z.string(),
+  score: z.number().optional(),
+});
+
+export const updateLeaderboardEntry = z.object({
+  name: z.string().optional(),
+  score: z.number().optional(),
+});
+
 export const eventOrganizerInfo = eventBaseInfo;
 export const eventAttendeeInfo = eventBaseInfo;
